fix: reorganize scripts
