+++ scripts/ceramic/clone_contribute_streams.py
@@ -3,7 +3,6 @@ from copy import deepcopy
 from streams import *
 import os
 import dotenv
-import json
 
 MAX_USERS_PER_COMMIT = 250
 MAX_TWEETS_PER_COMMIT = 100
