+++ scripts/ceramic/read_stream.py
