+++ scripts/ceramic/schemas/centaurs_stream_schema.json
