+++ scripts/ceramic/schemas/db_stream_schema.json
