+++ scripts/ceramic/schemas/default_centaurs_stream.json
