+++ scripts/ceramic/schemas/default_db_stream.json
