+++ scripts/ceramic/schemas/default_generic_points_stream.json
