+++ scripts/ceramic/schemas/generic_points_stream_schema.json
