+++ scripts/ceramic/write_schemas.py
