+++ scripts/contributors.json
